+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+## 0.3.4
+
+- Added `headers` option
+- Added `visitParams` option
+- Added `withCredentials` option
+
+## 0.3.3
+
+- Added `cookies` option
+
+## 0.3.2
+
+- Fixed module build
+
+## 0.3.1
+
+- Better handling of visit expiration in JavaScript
+- Wait for successful response to destroy `ahoy_track` cookie
+- Better compatibility with older browsers
+
+## 0.3.0
+
+- Removed jQuery dependency
+- Use `navigator.sendBeacon` by default in supported browsers
+- Removed support for IE < 9
+- Removed support for Bower
+
 ## 0.2.1
 
 - Fixed issue with duplicate events